third_party/WebKit/LayoutTests/inspector/tracing/timeline-network-resource-details-expected.txt

Tests the Timeline UI API for network requests.

Script resource loaded
URL: timeline-network-resource.js
Duration: string
Request Method: GET
Mime Type: string
URL: anImage.png
Duration: string
Request Method: GET
Mime Type: string
Preview: 
